backend-services/requirements.txt
flask
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary